model/requirements.txt
numpy==1.24.3
pandas==2.0.1
pip==23.1.2
py4j==0.10.9.7
pyspark==3.4.0
python-dateutil==2.8.2
pytz==2023.3
setuptools==67.7.2
six==1.16.0
tzdata==2023.3
wheel==0.40.0
kafka-python==2.0.2
flask==2.3.2